requirements.txt
aioredis==1.3.1
beautifulsoup4==4.9.0
bs4==0.0.1
Django==3.0.5
django-cors-headers==3.2.1
django-redis==4.11.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
hiredis==1.1.0
httplib2==0.9.2
idna==2.6
kavenegar==1.1.2
Pillow==5.1.0
postgres==3.0.0
protobuf==3.0.0
psycopg2-binary==2.8.5
psycopg2-pool==1.1
PyJWT==1.7.1
redis==3.4.1
simplejson==3.13.2
six==1.11.0
urllib3==1.22
wget==3.2

